% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCaNconstructorServer.R
\name{RCaNconstructorServer}
\alias{RCaNconstructorServer}
\title{RCaNconstructorServer
build the server of the Construtor}
\usage{
RCaNconstructorServer(input, output, session)
}
\arguments{
\item{input}{the shiny input}

\item{output}{the shiny output}

\item{session}{the shiny server}
}
\value{
a module server
}
\description{
RCaNconstructorServer
build the server of the Construtor
}
